+++ packages/iov-crypto/src/secp256k1.ts
@@ -1,18 +1,20 @@
-import { Encoding } from "@iov/encoding";
-
 import BN = require("bn.js");
 import elliptic = require("elliptic");
+import { As } from "type-tagger";
+
+import { Encoding } from "@iov/encoding";
 
 import { ExtendedSecp256k1Signature, Secp256k1Signature } from "./secp256k1signature";
 
 const secp256k1 = new elliptic.ec("secp256k1");
 const secp256k1N = new BN("FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFEBAAEDCE6AF48A03BBFD25E8CD0364141", "hex");
 
-export declare const Secp256k1KeypairSymbol: unique symbol;
-export type Secp256k1Keypair = typeof Secp256k1KeypairSymbol & {
+interface Keypair {
   readonly pubkey: Uint8Array;
   readonly privkey: Uint8Array;
-};
+}
+
+export type Secp256k1Keypair = Keypair & As<"secp256k1-keypair">;
 
 export class Secp256k1 {
   public static async makeKeypair(privkey: Uint8Array): Promise<Secp256k1Keypair> {
@@ -34,17 +36,15 @@ export class Secp256k1 {
       throw new Error("input data is not a valid secp256k1 private key");
     }
 
-    /* eslint-disable @typescript-eslint/no-object-literal-type-assertion */
-    // tslint:disable-next-line:no-object-literal-type-assertion
-    return {
+    const out: Keypair = {
       privkey: Encoding.fromHex(keypair.getPrivate("hex")),
       // encodes uncompressed as
       // - 1-byte prefix "04"
       // - 32-byte x coordinate
       // - 32-byte y coordinate
       pubkey: Encoding.fromHex(keypair.getPublic().encode("hex")),
-    } as Secp256k1Keypair;
-    /* eslint-enable @typescript-eslint/no-object-literal-type-assertion */
+    };
+    return out as Secp256k1Keypair;
   }
 
   // Creates a signature that is